refactor: moving methods to data and plotting files.
