+++ src/ansys/mapdl/core/reader/__init__.py
@@ -20,4 +20,9 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-from ansys.mapdl.core.reader.core import DPFResult
+from ansys.mapdl.core.reader.data import DPFResultData
+from ansys.mapdl.core.reader.plotting import DPFResultPlotting
+
+
+class DPFResult(DPFResultData, DPFResultPlotting):
+    """Unified class for accessing DPF result data and plotting"""
